requirements.txt
adafruit-circuitpython-neopixel==6.3.11